requirements.txt
fastapi==0.65.1 
uvicorn==0.11.3 
pydantic==1.8.2 
psycopg2-binary==2.8.6 